+++ src/components/Modal/index.js
@@ -1,13 +1,12 @@
 import style from './style.css'
 import ReactModal from 'react-modal'
 import { h, Component } from 'preact'
-import theme from '../Theme/style.css'
 import { getCSSMilisecsValue, wrapWithClass } from '../utils'
 
 const MODAL_ANIMATION_DURATION = getCSSMilisecsValue(style.modal_animation_duration)
 
 const WrapperContent = ({children}) =>
-  wrapWithClass(style.content, wrapWithClass(theme.step, children))
+  wrapWithClass(style.content, children)
 
 const Wrapper = ({children}) =>
   wrapWithClass(style.inner, <WrapperContent>{children}</WrapperContent>)
@@ -47,7 +46,7 @@ class ModalStrict extends Component {
     return (
       <ReactModal
         isOpen={this.state.isOpen || this.props.isOpen}
-        onRequestClose={this.onRequestClose}
+        onRequestClose={this.props.onRequestClose || this.onRequestClose}
         portalClassName={style.portal}
         overlayClassName={style.overlay}
         bodyClassName={style.modalBody}