+++ test/features/step_definitions/sdk_steps.rb
@@ -2,9 +2,15 @@ require_relative '../helpers/i18n_helper.rb'
 
 i18n = I18nHelper.new
 
-Given(/^I verify with (passport|identity_card|drivers_license)(?:| with (.+)?)$/) do |document_type, locale|
+Given(/^I initiate the verification process with(?: (.+)?)?$/) do |locale|
   i18n.load_locale(locale)
+  steps %Q{
+    Given I navigate to the SDK with "#{locale}"
+    Then I click on primary_button (SDK)
+  }
+end
 
+Given(/^I verify with (passport|identity_card|drivers_license)(?: with (.+)?)?$/) do |document_type, locale|
   if document_type == 'passport'
     key = 'capture.passport.front.title'
   elsif document_type == 'identity_card'
@@ -14,10 +20,10 @@ Given(/^I verify with (passport|identity_card|drivers_license)(?:| with (.+)?)$/
   end
 
   steps %Q{
-    Given I navigate to the SDK with "#{locale}"
-    When I click on primary_button (SDK)
+    Given I initiate the verification process with #{locale}
     Then I should see 3 document_select_buttons ()
     When I click on #{document_type} ()
+    Then I can confirm privacy terms
     Then page_title should include translation for "#{key}"
     And cross_device_header should include translation for "cross_device.switch_device.header"
   }
@@ -82,5 +88,18 @@ Then(/^I wait until (.*) has "([^"]*)"$/) do | page_element, key |
   text = i18n.translate(key)
   steps %Q{
     Then I wait until #{page_element} () contains "#{text}"
+  }
+end
+
+When(/^I press esc key$/) do
+  @driver.switch_to.active_element.send_keys(:escape)
+end
+
+
+Then(/^I can (confirm|decline) privacy terms$/) do | action |
+  next unless PRIVACY_FEATURE_ENABLED
+  steps %Q{
+    Then page_title should include translation for "privacy.title"
+    When I click on #{action}_privacy_terms ()
   }
 end